docs: explain custom ServiceCollection implementations
