+++ src/Java/myservicebus-rabbitmq/src/main/java/com/myservicebus/rabbitmq/RabbitMqFactoryConfigurator.java
@@ -123,7 +123,7 @@ public class RabbitMqFactoryConfigurator implements BusFactoryConfigurator {
 
     @Override
     public MessageBus build() {
-        ServiceCollection services = new ServiceCollection();
+        ServiceCollection services = ServiceCollection.create();
         configure(services);
         ServiceProvider provider = services.buildServiceProvider();
         return provider.getService(MessageBus.class);
