+++ src/Java/myservicebus-rabbitmq/src/test/java/com/myservicebus/rabbitmq/ServiceCollectionUsingTest.java
@@ -15,7 +15,7 @@ public class ServiceCollectionUsingTest {
 
     @Test
     public void addServiceBusWithUsingRegistersBus() {
-        ServiceCollection services = new ServiceCollection();
+        ServiceCollection services = ServiceCollection.create();
 
         services.from(MessageBusServices.class)
                 .addServiceBus(cfg -> {
